Uploading working version of Dearbhla's code.
TODO:
The delay in the flow function prevents the button press from being detected. It would be better to swap pin to one that supports hardware interrupts.

The first press lights up all leds - need to coordinate with Dearbhla to decide which ones will be underneath fibre optics (and so should be left on).